:sparkles: add new features and fix bugs
Features:

- Added ability to display your own images using upload button.
- Added ability to reset image view (zoom, offset, etc.) with '0' key.
- Added ability to toggle between smooth and pixelated image with 's' key.

Improvements:

- Changed how image is projected on canvas. Now at scale level 1:1 the
image is displayed in real size - 100px of the image take up 100px of
canvas space.

- Improved movement direction in image canvas. Now it's arguably more
intuitive when moving a zoomed in image. Moving left actually shifts
image to the right and vice versa.

- Improved background color in image canvas. No more yucky yellow.

Fixed bugs:

- Fixed floating point error in image rotation angle, which would
accumulate on each rotation. Now each full rotation should wrap around
2π, thus having constant minimal error.

- Fixed bug, where image would skew after rotation. This required
several changes.
First, the image aspect ratio is to be calculated as height over width.
This is because width is considered to be along the 'x' axis and so is
an independent value, but height is along the 'y' axis and is dependent
on the width.
Second, the image aspect ratio should be used to calculate texel
coordinate from vertex position. This is because the vertices use the
image aspect ratio to actually preserve it when displaying the image.
For example, a 450x300 image has 0.667 aspect ratio and so the quad's
top left point coordinates would be [-1, -0.677].

Misc:

- Added `ImageCanvas.fromImage` constructor to create image canvas
directly from image source.

- Added ability to specify per axis scaling factors in `Matrix.scaling`
method.

- Added check for 3D matrix before trying to convert it to homogenours
coordinates in `Matrix.toHomogenous` method.

- Added shader detaching after linking shader program in
`WebGLCustomRenderingContext.makeProgram` method.